Introduce auto-baseline parameter, move config to its own package
