+++ src/config/auto_baseline.py
@@ -1,0 +1,12 @@
+import datetime
+
+from pydantic import BaseModel
+
+
+class AutoBaselineConfig(BaseModel):
+    """
+    Configuration for the baseline metric.
+    """
+
+    enabled: bool
+    time: datetime.time
