+++ src/config/config.py
@@ -1,20 +1,19 @@
-from datetime import datetime
-
 import yaml
-from pydantic import BaseModel, field_validator
+from pydantic import BaseModel
 
 from typing import Any
-import emoji
 from src.autowiring.injectable import Injectable
+from src.config.auto_baseline import AutoBaselineConfig
+from src.config.config_metric import ConfigMetric
+
 from src.model.metric import Metric
 from src.model.notification import Notification
 
 
 class Configuration(BaseModel, Injectable):
-    metrics: list["ConfigMetric"]
+    metrics: list[ConfigMetric]
     notifications: list[Notification]
-    baseline: "BaselineConfig"
-    auto_baseline: bool = False
+    auto_baseline: AutoBaselineConfig
 
     def get_metrics(self) -> list[Metric]:
         return [Metric(**metric.model_dump()) for metric in self.metrics]
@@ -23,7 +22,7 @@ class Configuration(BaseModel, Injectable):
         return self.notifications
 
     def model_post_init(self, __context: Any) -> None:
-        if self.baseline.enabled:
+        if self.auto_baseline.enabled:
             # check that baselines are defined for all metrics
             for metric in self.metrics:
                 assert metric.baseline is not None
@@ -56,48 +55,3 @@ class ConfigurationProvider:
 
     def get_configuration(self) -> Configuration:
         return self.configuration
-
-
-class ConfigMetric(BaseModel):
-    """
-    Superclass of the Metric class to parse the configuration and perform post-initialization logic.
-    Technically, this could probably all be performed within the Metric class itself, but this separates concerns.
-    """
-
-    name: str
-    user_prompt: str
-    values: dict[str, int]
-    baseline: int | None = None
-    type: str | None = "enum"
-    emoji: bool = False
-
-    @field_validator("type")
-    @classmethod
-    def validate_type(cls, v):
-        if v is not None:
-            assert v in ["enum", "numeric"]
-        return v
-
-    def model_post_init(self, __context: Any) -> None:
-        if self.type == "numeric":
-            assert "lower_bound" in self.values
-            assert "upper_bound" in self.values
-            assert self.emoji is False
-            self.values = {
-                str(i): i
-                for i in range(
-                    self.values["lower_bound"], self.values["upper_bound"] + 1
-                )
-            }
-
-        if self.emoji:
-            self.values = {emoji.emojize(k): v for k, v in self.values.items()}
-
-
-class BaselineConfig(BaseModel):
-    """
-    Configuration for the baseline metric.
-    """
-
-    enabled: bool
-    time: datetime.time
