+++ test/resources/config.test.yaml
@@ -23,6 +23,6 @@ notifications:
   - text: "It's time to record your mood!"
     time: "18:00"
 
-baseline:
+auto_baseline:
   enabled: true
-  auto: false
+  time: "23:00"
